Initialize maxmove variable for all file types.
maxmove sets the number of molecules that may be moved to a better
position and defaults to all molecules in the structure section. However
this was only being parsed for the pdb file format, resulting in 0
molecules being moved in movebad in heuristics.F90 and failure to
converge for inputs of non-pdb file formats.

This PR sets the default value regardless of file type and parses it out
of the structure section.