reads input data
NOTES
packmol either reads from standard input or from the specified
input file

if the name of the input-file is an empty string packmol assumes
that the user has not provided any command-line arguments and thus it
fallsback to the default workflow of reading data from standard input;
otherwise we open the input-file and read the data from it. After
we are done we close the input file.