handles the output filename according to the cli
NOTES
If the user has provided a filename for the output file via the
command-line we use that; otherwise we fallback to the default
workflow of fecthing the filename from the input file.

Note that we did not apply indentation to make the git-diff easier
to read.

EXAMPLE

```sh
packmol -in input.txt -out output.txt
```

we get this

```

 PACKMOL - Packing optimization for the automated generation of
 starting configurations for molecular dynamics simulations.

                                                             Version 21.0.4

 input: input.txt
 output: output.txt
  Packmol must be run with: packmol < inputfile.inp

  Userguide at: http://m3g.iqm.unicamp.br/packmol

  Reading input file... (Control-C aborts)
  Types of coordinate files specified: pdb
  Seed for random number generator:      1234567
  Output file: output.txt
```

when we invoke packmol in the usual way, that is via:

```sh
packmol < input.txt
```

```

 PACKMOL - Packing optimization for the automated generation of
 starting configurations for molecular dynamics simulations.

                                                             Version 21.0.4

  Packmol must be run with: packmol < inputfile.inp

  Userguide at: http://m3g.iqm.unicamp.br/packmol

  Reading input file... (Control-C aborts)
  Types of coordinate files specified: pdb
  Seed for random number generator:      1234567
  Output file: system.pdb
```

the input file for this example was this one:

```
tolerance 2.0

filetype pdb

output system.pdb

structure data/IRMOF-1.pdb
  number 1
  inside box 0. 0. 0. 40 40 40
end structure

structure data/EtOH.pdb
  number 20
  inside box 0. 0. 0. 40 40 40
end structure
```

note that the name of the output file is set according to the
workflow (default or cli-mode)